+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.5.7
+- Support reading outputs from CellRanger version 3 and later in `read_tenx_mtx`.
+  - Handle both `genes.csv` and `features.csv` for gene annotations.
+
 ## Version 0.5.6
 
 - Check if the column names of the alternative experiments match with the column names of the main experiment. This is the equivalent to the ``withDimnames`` parameter in the R implementation.